fix(#3552): split arrow_size.rs by responsibility + two accidental doc lists
file-size ratchet: my change took arrow_size.rs from 657 (main) to 813,
over the 800 limit. Doctrine for a file THIS diff pushed over is to split,
not to set CQLITE_ALLOW_FILE_GROWTH -- that env var was right for #3741's
already-oversized compression.rs and would be a waiver against my own diff
here.

Seam: per-SLOT charging stays in arrow_size.rs (constants, public entry
points, Estimator); per-ROW iteration over a RESOLVED column set moves to
arrow_size_prepared.rs -- the shared charge_row loop plus PreparedColumns and
the two cell resolutions that map the accumulator's storage layouts onto the
(shape, cell) pairs charge_row consumes. Nothing in the child knows how a
slot is charged; nothing in the parent knows how a caller stores cells.
Wired as a #[path] child module exactly like arrow_size_shape.rs /
arrow_size_render.rs, and PreparedColumns is re-exported from arrow_size at
its unchanged pub(in crate::export) visibility, so the accumulator's import
path does not change and nothing widened. 813 -> 674 + 196; no lines shaved
to land under the limit.

Also fixed en route: round 3 inserted the two cell-resolution helpers using
an anchor inside PreparedColumns' doc comment, which left PreparedColumns'
opening paragraph attached to staged_cell. Restored to its owner.

clippy doc_lazy_continuation (8 hits): my round-2 rewrap put a '+' at the
start of a doc line in StreamSubPhase::EncodeFraming, which markdown reads
as a list item and every following line as a lazy continuation. Rewrapped,
and swept every touched file for the same shape -- found one more in the new
sparse_cell doc (also mine, also fixed). The remaining '-'/'+' line starts
are pre-existing and inside code fences or real lists.

cargo fmt --all -- --check: exit 0.
